+++ python/sidecar/search/service.py
@@ -6,9 +6,9 @@ retrieves information about a set of papers.
 import logging
 
 from semanticscholar import SemanticScholar
-
-from .stopwords import stopwords
-from .typing import ResponseStatus, S2SearchResult, SearchRequest, SearchResponse
+from sidecar.search.constants import stopwords
+from sidecar.search.schemas import S2SearchResult, SearchRequest, SearchResponse
+from sidecar.typing import ResponseStatus
 
 logger = logging.getLogger(__name__)
